make vegafusion-runtime not depend on full datafusion crate
