+++ vegafusion-datafusion-udfs/src/udfs/mod.rs
@@ -1,3 +1,5 @@
 pub mod array;
 pub mod datetime;
 pub mod math;
+pub mod member;
+pub mod object;
