+++ vegafusion-runtime/src/expression/compiler/builtin_functions/type_coercion/to_boolean.rs
@@ -1,6 +1,6 @@
 use crate::expression::compiler::utils::to_boolean;
-use datafusion::common::DFSchema;
-use datafusion::logical_expr::Expr;
+use datafusion_expr::Expr;
+use vegafusion_common::datafusion_common::DFSchema;
 use vegafusion_core::error::{Result, VegaFusionError};
 
 pub fn to_boolean_transform(args: &[Expr], schema: &DFSchema) -> Result<Expr> {
