+++ vegafusion-runtime/src/expression/compiler/builtin_functions/type_coercion/to_number.rs
@@ -1,7 +1,7 @@
 use crate::expression::compiler::utils::{cast_to, is_numeric_datatype};
-use datafusion::arrow::datatypes::DataType;
-use datafusion::common::DFSchema;
-use datafusion::logical_expr::{Expr, ExprSchemable};
+use datafusion_expr::{Expr, ExprSchemable};
+use vegafusion_common::arrow::datatypes::DataType;
+use vegafusion_common::datafusion_common::DFSchema;
 use vegafusion_core::error::{Result, ResultWithContext, VegaFusionError};
 
 pub fn to_number_transform(args: &[Expr], schema: &DFSchema) -> Result<Expr> {
