+++ vegafusion-runtime/src/expression/compiler/call.rs
@@ -7,16 +7,14 @@ use crate::expression::compiler::builtin_functions::type_checking::isvalid::is_v
 use crate::expression::compiler::compile;
 use crate::expression::compiler::config::CompilationConfig;
 use crate::expression::compiler::utils::cast_to;
-use datafusion::arrow::datatypes::DataType;
-use datafusion::common::DFSchema;
-use datafusion::logical_expr::Expr;
-use datafusion::physical_plan::udf::ScalarUDF;
-use datafusion_expr::BuiltinScalarFunction;
+use datafusion_expr::{BuiltinScalarFunction, Expr, ScalarUDF};
 use std::collections::HashMap;
 use std::ops::Deref;
 use std::str::FromStr;
 use std::sync::Arc;
+use vegafusion_common::arrow::datatypes::DataType;
 use vegafusion_common::data::table::VegaFusionTable;
+use vegafusion_common::datafusion_common::DFSchema;
 use vegafusion_common::error::{Result, ResultWithContext, VegaFusionError};
 use vegafusion_core::proto::gen::expression::{
     expression, literal, CallExpression, Expression, Literal,
