+++ vegafusion-runtime/src/expression/compiler/config.rs
@@ -1,9 +1,9 @@
 use crate::expression::compiler::call::{default_callables, VegaFusionCallable};
 use crate::task_graph::timezone::RuntimeTzConfig;
-use datafusion::scalar::ScalarValue;
 use num_traits::float::FloatConst;
 use std::collections::HashMap;
 use vegafusion_common::data::table::VegaFusionTable;
+use vegafusion_common::datafusion_common::ScalarValue;
 
 #[derive(Clone)]
 pub struct CompilationConfig {
