+++ vegafusion-runtime/tests/test_transform_formula.rs
@@ -2,7 +2,8 @@
 extern crate lazy_static;
 
 mod util;
-use datafusion::scalar::ScalarValue;
+
+use datafusion_common::ScalarValue;
 use util::check::check_transform_evaluation;
 use util::datasets::vega_json_dataset;
 use vegafusion_core::spec::transform::formula::FormulaTransformSpec;
